requirements.txt
requests
pyyaml
beautifulsoup4
PyExecJS